requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
sqlmodel==0.0.16
httpx==0.28.1
python-dotenv==1.0.1
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
firebase-admin==6.9.0
